Evaluation Works but Int to Str Conversion Has an Error

+++ C/LinuxProjects/Calculator/MainFiles/conversions.c
@@ -9,7 +9,7 @@ int strToint(char* string){
 	//for each char turn into int
 	int converted = 0;
 	for(int i = 0; string[i] != '\0'; i++){
-		converted += (string[i]-48) * pow(10,i);
+		converted += (string[i]-48) * (int)pow(10,i);
 	}
 	return converted;
 }
@@ -19,13 +19,13 @@ char* intTostr(int invalue, int base){
 	//iterate by base^i to get each value
 	//get digits for size of char
 	int digits = 0;
-	for(digits = 0; invalue/(pow(base,digits)) != 0; digits++)
+	for(digits = 0; invalue/((int)pow(base,digits)) != 0; digits++)
 		;
 
 	char* converted = (char*)malloc(sizeof(char) * (digits+1));
 	int difference = 0;
 	for(int i = 0; i < digits; i++){
-		int power = pow(base,(digits-1)-i);
+		int power = (int)pow(base,(digits-1)-i);
 		converted[i] = ((invalue-difference)/power) -48; //to make sure its the int
 		difference += (converted[i]+48) * base;
 	}
